feat: refactor add shared files
