+++ integration_tests/specs/shadcn/field.ts
@@ -1,6 +1,6 @@
 import React from 'react';
-import { FieldFixture } from './shadcn-component';
-import { runShadcnCase } from './shadcn-test-utils';
+import { FieldFixture } from './shared/shadcn-component';
+import { runShadcnCase } from './shared/shadcn-test-utils';
 
 describe('Shadcn field integration', () => {
   it('shadcn_field', async () => {
