+++ integration_tests/specs/shadcn/progress.ts
@@ -1,6 +1,6 @@
 import React from 'react';
-import { ProgressFixture } from './shadcn-component';
-import { runShadcnCase } from './shadcn-test-utils';
+import { ProgressFixture } from './shared/shadcn-component';
+import { runShadcnCase } from './shared/shadcn-test-utils';
 
 describe('Shadcn progress integration', () => {
   it('shadcn_progress', async () => {
