+++ integration_tests/specs/shadcn/shared/shadcn-component.css
